add apiplat, configure entites for serialization & deserialization, add jwt token
