+++ config/packages/api_platform.yaml
@@ -1,5 +1,5 @@
 api_platform:
-    title: Hello API Platform
+    title: API Bilemo
     version: 1.0.0
     formats:
         jsonld: ['application/ld+json']
