+++ src/Entity/Phone.php
@@ -6,32 +6,42 @@ use ApiPlatform\Metadata\ApiResource;
 use App\Repository\PhoneRepository;
 use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
+use Symfony\Component\Serializer\Attribute\Groups;
 
 #[ORM\Entity(repositoryClass: PhoneRepository::class)]
-#[ApiResource]
+#[ApiResource(
+    normalizationContext: ["groups" => "phones:read"]
+)]
 class Phone
 {
     #[ORM\Id]
     #[ORM\GeneratedValue]
     #[ORM\Column]
+    #[Groups(["phones:read"])]
     private ?int $id = null;
 
     #[ORM\Column(length: 255)]
+    #[Groups(["phones:read"])]
     private ?string $name = null;
 
     #[ORM\Column(type: Types::TEXT)]
+    #[Groups(["phones:read"])]
     private ?string $content = null;
 
     #[ORM\Column(length: 255)]
+    #[Groups(["phones:read"])]
     private ?string $color = null;
 
     #[ORM\Column]
+    #[Groups(["phones:read"])]
     private ?int $storage = null;
 
     #[ORM\Column]
+    #[Groups(["phones:read"])]
     private ?float $price = null;
 
     #[ORM\Column(length: 255)]
+    #[Groups(["phones:read"])]
     private ?string $image = null;
 
     public function getId(): ?int
