+++ src/Events/PasswordEncoderSubscriber.php
@@ -1,0 +1,37 @@
+<?php
+
+namespace App\Events;
+
+use ApiPlatform\Symfony\EventListener\EventPriorities;
+use App\Entity\User;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Symfony\Component\HttpKernel\Event\ViewEvent;
+use Symfony\Component\HttpKernel\KernelEvents;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
+
+readonly class PasswordEncoderSubscriber implements EventSubscriberInterface
+{
+    public function __construct(
+        private UserPasswordHasherInterface $hasher,
+    ) {
+    }
+
+    public static function getSubscribedEvents(): array
+    {
+        return [
+            KernelEvents::VIEW => ['encodePassword', EventPriorities::PRE_WRITE]
+        ];
+    }
+
+    public function encodePassword(ViewEvent $event): void
+    {
+
+        $result = $event->getControllerResult();
+        $method = $event->getRequest()->getMethod();
+
+        if($result instanceof User && $method === "POST") {
+            $hash = $this->hasher->hashPassword($result, $result->getPassword());
+            $result->setPassword($hash);
+        }
+    }
+}
